+++ tcex/tcex_batch_v2.py
@@ -2,6 +2,7 @@
 """ThreatConnect Batch Import Module"""
 import hashlib
 import json
+import math
 import os
 import re
 import shelve
@@ -77,7 +78,9 @@ class TcExBatch(object):
         self._batch_count = 0
 
         # default properties
-        self._poll_interval = 15
+        self._batch_data_count = None
+        self._poll_interval = None
+        self._poll_interval_times = []
         self._poll_timeout = 3600
 
         # containers
@@ -915,7 +918,7 @@ class TcExBatch(object):
         indicator_obj = Mutex(mutex, rating, confidence, xid)
         return self._indicator(indicator_obj)
 
-    def poll(self, batch_id, interval=None, timeout=None, halt_on_error=True):
+    def poll(self, batch_id, retry_seconds=None, back_off=None, timeout=None, halt_on_error=True):
         """Poll Batch status to ThreatConnect API.
 
         .. code-block:: javascript
@@ -946,10 +949,28 @@ class TcExBatch(object):
         if self.halt_on_poll_error is not None:
             halt_on_error = self.halt_on_poll_error
 
-        if interval is None:
-            interval = self.poll_interval
+        # initial poll interval
+        if self._poll_interval is None and self._batch_data_count is not None:
+            # calculate poll_interval base off the number of entries in the batch data
+            # with a minimum value of 5 seconds.
+            self._poll_interval = max(math.ceil(self._batch_data_count / 250), 5)
+        elif self._poll_interval is None:
+            # if not able to calculate poll_interval default to 15 seconds
+            self._poll_interval = 15
+
+        # poll retry back_off factor
+        if back_off is None:
+            poll_interval_back_off = 2.5
         else:
-            interval = int(interval)
+            poll_interval_back_off = float(batch_id)
+
+        # poll retry seconds
+        if retry_seconds is None:
+            poll_retry_seconds = 5
+        else:
+            poll_retry_seconds = int(retry_seconds)
+
+        # poll timeout
         if timeout is None:
             timeout = self.poll_timeout
         else:
@@ -958,10 +979,15 @@ class TcExBatch(object):
             'includeAdditional': 'true'
         }
 
-        poll_time = 0
+        poll_count = 0
+        poll_time_total = 0
         data = {}
         while True:
+            poll_count += 1
+            poll_time_total += self._poll_interval
+            time.sleep(self._poll_interval)
             try:
+                # retrieve job status
                 r = self.tcex.session.get('/v2/batch/{}'.format(batch_id), params=params)
                 if not r.ok or 'application/json' not in r.headers.get('content-type', ''):
                     self.tcex.handle_error(545, [r.status_code, r.text], halt_on_error)
@@ -969,28 +995,31 @@ class TcExBatch(object):
                 data = r.json()
                 if data.get('status') != 'Success':
                     self.tcex.handle_error(545, [r.status_code, r.text], halt_on_error)
-                if data.get('data', {}).get('batchStatus', {}).get('status') == 'Completed':
-                    self.tcex.log.debug('Batch Status: {}'.format(data))
-                    return data
-                time.sleep(interval)
             except Exception as e:
                 self.tcex.handle_error(540, [e], halt_on_error)
 
+            if data.get('data', {}).get('batchStatus', {}).get('status') == 'Completed':
+                # store last 5 poll times to use in calculating average poll time
+                self._poll_interval_times = self._poll_interval_times[-4:] + [poll_time_total]
+                # new poll interval is average of last 5 poll times
+                self._poll_interval = (
+                    math.ceil(sum(self._poll_interval_times) / len(self._poll_interval_times)))
+
+                if poll_count == 1:
+                    # if completed on first poll, reduce poll interval.
+                    self._poll_interval -= .5
+
+                self.tcex.log.debug('Batch Status: {}'.format(data))
+                return data
+
+            # update poll_interval for retry
+            self._poll_interval = poll_retry_seconds + int(poll_count * poll_interval_back_off)
+            self.tcex.log.debug('Batch poll time: {} seconds'.format(poll_time_total))
+
             # time out poll to prevent App running indefinitely
-            if poll_time >= timeout:
+            if poll_time_total >= timeout:
                 self.tcex.handle_error(550, [timeout], True)
-            poll_time += interval
-            self.tcex.log.debug('Batch poll time: {} seconds'.format(poll_time))
-
-    @property
-    def poll_interval(self):
-        """Return current poll interval value."""
-        return self._poll_interval
-
-    @poll_interval.setter
-    def poll_interval(self, interval):
-        """Set the poll interval value."""
-        self._poll_interval = int(interval)
+            self.tcex.log.info('Batch poll time: {} seconds'.format(poll_time_total))
 
     @property
     def poll_timeout(self):
@@ -1253,6 +1282,8 @@ class TcExBatch(object):
             halt_on_error = self.halt_on_batch_error
 
         content = self.data
+        # store the length of the batch data to use for poll interval calculations
+        self._batch_data_count = len(content)
         if self._debug:
             # special code for debugging App using batchV2.
             self.write_batch_json(content)
@@ -1284,6 +1315,8 @@ class TcExBatch(object):
             halt_on_error = self.halt_on_batch_error
 
         content = self.data
+        # store the length of the batch data to use for poll interval calculations
+        self._batch_data_count = len(content)
         if content.get('group') or content.get('indicator'):
             headers = {'Content-Type': 'application/octet-stream'}
             try: